Create README.md
The insights gained from this analysis would help the organization develop effective strategies to enhance employee relationship, creating more stable and  productive environment and ultimately creating a more stable and productive work environment. This project was not only valuable for reducing turnover costs of employees but also for fostering a relationship between the  employees and its owner thereby creating room for more growth and development.